move importdata to its own file
